// src/app/models/job-card-filter.ts
export class JobCardFilter {
    WorkshopID: number;
    CustomerName: string;
    CustomerMobile: string;
    PlateNumber: string;
    UserID: number;
    EmployeeID: number;
    VIN: string;
    JobCardNumber: string;
    CheckInDate: any;
    CheckOutDate: any;
    MakeID: number;
    ModelID: number;
    StartRow: number;
    RowsPerPage: number;
    UserRoleID: number;
    InvoiceStatusID: number;
    pageNumber = 1;
    queuePageNumber = 1;
    TabID: number;
    SearchQuery: string;
    ProductionYear: number;

}